backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=24.1.0
httpx>=0.27.0

# PII Detection
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# File Parsers
python-docx>=1.1.0
openpyxl>=3.1.0
extract-msg>=0.48.0

# AI Providers
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
PyJWT[crypto]>=2.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
